[iOS] MainCellView + TokenCellView 통합 완료
- TokenCellView+SubViews 파일 추가
    - TokenCellView의 서브뷰를 분리하였다.
- 프로그레스바, 초 Text 추가
- MainCellView 삭제
